refactor: move twitch IRC infra into twitch/ module
Groups tls, token_storage, setup (twitch_setup), seventv, and all
handlers into src/twitch/ for clearer module boundaries.
